+++ utilities/tools.descartes.teastore.kieker.probes/src/main/java/tools/descartes/teastore/kieker/probes/records/IPayloadCharacterization.java
@@ -1,4 +1,4 @@
-package tools.descartes.teastore.kieker.events;
+package tools.descartes.teastore.kieker.probes.records;
 
 import kieker.common.record.IMonitoringRecord;
